feat(attendance): overview page for attendance register
explaining different options and avoiding an immediate loading of lots of data

(hopefully) closes #487
