+++ src/app/child-dev-project/attendance/attendance-analysis/attendance-analysis.component.scss
@@ -1,0 +1,19 @@
+.attendance-line {
+  margin: 10px;
+}
+
+.table-list{
+  margin-top: 10px;
+}
+
+.table-list td {
+  padding-left: 8px;
+}
+
+.table-list-item:hover{
+  background: #f5f5f5;
+}
+
+.filter-panel {
+  padding-top: 10px;
+}
